Fixes bug with db seed user passwords
Added a static schema method to the User schema to generate a random
passphrase, that passes the owasp strength test.

Updated the db seed configuration to use a randomly generated passphrase
from the new User schema method, as the passwords for the seeded users.

Added the generate-password package to the project.