fix(users): User email to accept null regardless of unique
Provider Twitter (refer twitter.js file) is not producing an email
address on callback. When two different users signs up using twitter the
second user will not be able to sign up if the first user didn't set the
email address manually. This is due to the error of email already exist,
in actual terms this is the null value which is conflicting the unique
constraint.