feat(articles): Mongoose Filtering & Pagination
Adds functionality for filtering, sorting, and paging, mongoose queries
on the server-side.

Added error message to list view when server returns error

NOTE: This has been implemented for just the main (public) Articles
list view. We can discuss if we want this on the Admin list as well.

TODO: (Are these necessary for the framework? Or should we let
our users roll their own client-side filtering mechanisms?)
1) Add client-side search input for filtering
2) Create directive for search inputs

Closes meanjs/mean#1048