feat(gulp): test:server:watch task onlyChanged
Sets the behavior of the test:server:watch Gulp task to only run a single
test suite if the modified server file is a test suite. Previously this was
acheived through the use of the `onlyChanged` parameter. The parameter has
been removed since this is almost certainly the only use case.

Also, refactors this Gulp task to be a stand-alone watch task rather than
including it in the 'watch' task.

Removed the need for using yargs library, and removed the yargs package
from the framework.

Closes #1297