fix(articles): Article controllers name conflicts
Fixes the naming conflicts for the Articles controllers.

Due to how Angular injects the controllers into the StateProvider,
naming conflicts were caused between the Articles public & admin module
controllers.

To resolve the issue the referenced controllers in the Articles admin
route configurations must be unique, and match up with the Admin
controllers.