Enhanced the MongooseFiltering service
Refactored the MongooseFiltering service to be more "class-like".

Modified the existing `list()` method of the articles server controller
to allow paging options to be sent as query-string parameters. This will
allow any application to make a GET request to the */api/articles*
route, and get back a paged list of Articles as a response.