feat(core): Add manifest.json
This PR adds a JSON file which provides information about the app,
offering a quicker access and a richer experience for the users.

Not sure if this should be added right away since it is still in development,
however since it seems to be considered a best practice to have one, here it is.