fix(gulp): Gulp tasks override NODE_ENV
Fixes an issue with the "defualt" & "build" Gulp tasks that override the current NODE_ENV. These two tasks should inherit the NODE_ENV that has been set.

Also, adds a sequence to the "prod" Gulp task to ensure the NODE_ENV is set to "prod" before the "build" task is run.